Refactor ToString tests
